requirements.txt
streamlit == 1.30.0
pandas
gspread
oauth2client
langchain>=0.0.217
openai>=1.2
duckduckgo-search
anthropic>=0.3.0
trubrics>=1.4.3
streamlit-feedback
langchain-community
st-gsheets-connection
Pillow